Refactor package and remove 'minecraft' prefix from definitions
Also add data-generated sounds.json
