+++ src/generated_test/resources/.cache/cache
@@ -49,6 +49,7 @@ db2cd929128c8a6e8e0b02437f90ad8d24238206 assets/data_gen_test/models/block/torch
 098d8b38409fff6cc14a4e56355bedb1fc5931d1 assets/data_gen_test/models/item/test_block.json
 b029993275f16569b0edf093362f55df6ea6f9be assets/data_gen_test/models/item/test_block_model.json
 726e1884e53151492ca05bd752c67031603570f6 assets/data_gen_test/models/item/test_generated_model.json
+61e57e44de5506da3937499a4fe46f84b856ba52 assets/data_gen_test/sounds.json
 1767c874758bd7c3235db446b6219e6e1edd25b4 assets/minecraft/blockstates/acacia_door.json
 e41383b4c13ec922c0f5cfa2cc313d88fd294685 assets/minecraft/blockstates/acacia_fence.json
 967d268bfcf32474b61578bef16e853b6360dd73 assets/minecraft/blockstates/acacia_fence_gate.json
