+++ src/main/java/net/minecraftforge/common/data/SoundDefinitionsProvider.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package net.minecraftforge.client.sounds.generators;
+package net.minecraftforge.common.data;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
